MenuPermission, Menu, UserAccountGroup, UserGroup ile application command, validator,service ,Inf.Katmanı service ler webapi controller, conf.servicler yazıldı yarın WeBapi yi ayağa kaldırmaya çalışacağım

+++ VelorusNet8.Application/VelorusNet8.Application/Commands/Group/DeleteUserGroupCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using MediatR;
+﻿using FluentValidation;
+using MediatR;
 using VelorusNet8.Application.Interface.Group;
 using VelorusNet8.Application.Interface.Menus;
 
@@ -9,19 +10,29 @@ public class DeleteUserGroupCommandHandler : IRequestHandler<DeleteUserGroupComm
     private readonly IUserGroupRepository _userGroupRepository;
     private readonly IUserAccountGroupRepository _userAccountGroupRepository;
     private readonly IMenuPermissionRepository _menuPermissionRepository;
+    private readonly IValidator<DeleteUserGroupCommand> _validator;
 
     public DeleteUserGroupCommandHandler(
         IUserGroupRepository userGroupRepository,
         IUserAccountGroupRepository userAccountGroupRepository,
-        IMenuPermissionRepository menuPermissionRepository)
+        IMenuPermissionRepository menuPermissionRepository,
+        IValidator<DeleteUserGroupCommand> validator)
     {
         _userGroupRepository = userGroupRepository;
         _userAccountGroupRepository = userAccountGroupRepository;
         _menuPermissionRepository = menuPermissionRepository;
+        _validator = validator;
     }
 
     public async Task<Unit> Handle(DeleteUserGroupCommand request, CancellationToken cancellationToken)
     {
+        // Doğrulama işlemi
+        var validationResult = await _validator.ValidateAsync(request, cancellationToken);
+
+        if (!validationResult.IsValid)
+        {
+            throw new ValidationException(validationResult.Errors);
+        }
         // Önce UserGroup ile ilgili tüm ilişkileri temizliyoruz.
         await _userAccountGroupRepository.RemoveByGroupIdAsync(request.Id, cancellationToken);
         var menuPermissions = await _menuPermissionRepository.GetByUserGroupIdAsync(request.Id, cancellationToken);
